# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Red Fox Hotel HRMS & Payroll API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /branches:
    get:
      operationId: getBranches
      tags: [branches]
      summary: List all branches
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # DEPARTMENTS
  /departments:
    get:
      operationId: getDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # SHIFTS
...
  /continue-duties:
    get:
      operationId: getContinueDuties
      tags: [continueDuty]
      summary: List continue duty entries
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContinueDuty"
    post:
      operationId: createContinueDuty
      tags: [continueDuty]
      summary: Create a continue duty entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContinueDutyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinueDuty"
  /continue-duties/{id}:
    get:
      operationId: getContinueDuty
      tags: [continueDuty]
      summary: Get a continue duty entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinueDuty"
    patch:
      operationId: updateContinueDuty
      tags: [continueDuty]
      summary: Update a continue duty entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContinueDutyUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinueDuty"
    delete:
      operationId: deleteContinueDuty
      tags: [continueDuty]
      summary: Delete a continue duty entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # PAYROLL
...
components:
  schemas:
...
    Branch:
      type: object
      required: [id, name, address, phone, email]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        employeeCount:
          type: integer
        payrollCost:
          type: number
        createdAt:
          type: string
    BranchInput:
      type: object
      required: [name, address, phone, email]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        managerId:
          type: ["integer", "null"]
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        managerId:
          type: ["integer", "null"]
    Department:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeCount:
          type: integer
        createdAt:
          type: string
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
...
    ContinueDuty:
      type: object
      required: [id, employeeId, date, amount]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        date:
          type: string
        amount:
          type: number
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    ContinueDutyInput:
      type: object
      required: [employeeId, date, amount]
      properties:
        employeeId:
          type: integer
        date:
          type: string
        amount:
          type: number
        remarks:
          type: ["string", "null"]
    ContinueDutyUpdate:
      type: object
      properties:
        date:
          type: string
        amount:
          type: number
        remarks:
          type: ["string", "null"]
...